+++ src/main/java/terminator/command/CommandParser.java
@@ -1,4 +1,4 @@
-package command;
+package terminator.command;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
