+++ src/main/java/terminator/command/DeleteCommand.java
@@ -1,33 +1,32 @@
-package command;
+package terminator.command;
 
-import task.Task;
+import terminator.task.Task;
 
 import java.util.ArrayList;
 
-public class UnmarkCommand extends Command {
+public class DeleteCommand extends Command {
 
     private final String input;
 
-    public UnmarkCommand(String input) {
+    public DeleteCommand(String input) {
         this.input = input;
     }
 
     private static final String ERR_MSG = """
-            Index to unmark cannot be empty.
+            Index to delete cannot be empty.
             
-            Usage: unmark <index>""";
-
+            Usage: delete <index>""";
     @Override
     public void execute(ArrayList<Task> todoList) throws DukeException {
-        if (input == null) throw new DukeException(ERR_MSG);
-        String trimmedInput = input.trim();
+        if (input == null) {
+            throw new DukeException(ERR_MSG);
+        }
         try {
-            int idx = Integer.parseInt(trimmedInput) - 1;
+            int idx = Integer.parseInt(input.trim()) - 1;
 
             if (idx >= 0 && idx < todoList.size()) {
-                System.out.println("Objective reopened:");
-                Task currentTask = todoList.get(idx);
-                currentTask.markAsIncomplete();
+                System.out.println("Successfully wiped all traces of the task from the database.");
+                Task currentTask = todoList.remove(idx);
                 System.out.println(currentTask);
             } else {
                 throw new DukeException("Index out of bounds.");
