+++ src/main/java/terminator/command/DukeException.java
@@ -1,4 +1,4 @@
-package command;
+package terminator.command;
 
 public class DukeException extends Exception {
     public DukeException(String msg) {
