+++ src/main/java/terminator/command/UnmarkCommand.java
@@ -1,32 +1,33 @@
-package command;
+package terminator.command;
 
-import task.Task;
+import terminator.task.Task;
 
 import java.util.ArrayList;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
 
     private final String input;
 
-    public DeleteCommand(String input) {
+    public UnmarkCommand(String input) {
         this.input = input;
     }
 
     private static final String ERR_MSG = """
-            Index to delete cannot be empty.
+            Index to unmark cannot be empty.
             
-            Usage: delete <index>""";
+            Usage: unmark <index>""";
+
     @Override
     public void execute(ArrayList<Task> todoList) throws DukeException {
-        if (input == null) {
-            throw new DukeException(ERR_MSG);
-        }
+        if (input == null) throw new DukeException(ERR_MSG);
+        String trimmedInput = input.trim();
         try {
-            int idx = Integer.parseInt(input.trim()) - 1;
+            int idx = Integer.parseInt(trimmedInput) - 1;
 
             if (idx >= 0 && idx < todoList.size()) {
-                System.out.println("Successfully wiped all traces of the task from the database.");
-                Task currentTask = todoList.remove(idx);
+                System.out.println("Objective reopened:");
+                Task currentTask = todoList.get(idx);
+                currentTask.markAsIncomplete();
                 System.out.println(currentTask);
             } else {
                 throw new DukeException("Index out of bounds.");
