+++ src/main/java/terminator/task/DeadlineTask.java
@@ -1,4 +1,4 @@
-package task;
+package terminator.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
