+++ src/main/java/terminator/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package terminator.task;
 
 public abstract class Task {
     protected String description;
