+++ src/main/java/terminator/task/TodoTask.java
@@ -1,4 +1,4 @@
-package task;
+package terminator.task;
 
 public class TodoTask extends Task {
     public TodoTask(String description) {
